Changed type od id parameter to prevent implicit conversion (#14)

+++ shm/inc/ship.hpp
@@ -6,9 +6,9 @@ class Ship {
 public:
     Ship()
         : id_(-1) {}
-    Ship(int capacity, int maxCrew, int speed, const std::string& name, size_t id)
+    Ship(int capacity, int maxCrew, int speed, const std::string& name, unsigned int id)
         : capacity_(capacity), maxCrew_(maxCrew), crew_(0), speed_(speed), name_(name), id_(id) {}
-    Ship(int maxCrew, int speed, size_t id)
+    Ship(int maxCrew, int speed, unsigned int id)
         : Ship(0, maxCrew, speed, "", id) {}
 
     void setName(const std::string& name) { name_ = name; }
